Remove the temporal bounds override feature
`pond set-temporal-bounds` did not work, and using it made things worse:
applying bounds to a series made every subsequent query on that file fail
outright, with no way to undo it.

The filter wrapper hard-required a `Timestamp(Second, tz)` column, but real
series carry millisecond timestamps, so the moment an override existed the
scan returned `Expected timestamp column with second precision` -- for `cat`,
for `describe`, for everything. There was no clear/inspect surface: `describe`
never showed overrides, and `set-temporal-bounds` with no arguments errored
instead of removing them, so a file in that state stayed broken.

Beyond the hard failure the design never held together:

- Units disagreed. The CLI parsed to milliseconds, `min/max_event_time` were
  microseconds, and the scan divided by 1000 to get seconds under a comment
  saying "for HydroVu data". `effective_temporal_range()` mixed the microsecond
  range with millisecond overrides -- a 1000x error, harmless only because
  nothing called it. temporal.rs itself opened by admitting the millisecond
  choice "is arbitrary, and not correct".
- One-sided bounds were silently dropped. The CLI accepted `--min-time` alone,
  but both `temporal_overrides()` accessors returned `Some` only when both
  fields were set, so the value was written to the oplog and never read back.
- Enforcement was conditional: bounds applied only when `additional_urls` was
  empty, so globbed and multi-file reads of the same file skipped them.
- A second, unused implementation sat in tlogfs::query::temporal_filter with no
  production callers, including a predicate stub that always returned true.
- The representation fought storage. Bounds rode on an appended zero-byte
  version, which forced `inherited_overrides()` and a bespoke live-row resolver
  into the collapse path purely to stop collapse from deleting them.

`detect-overlaps`, the companion inspection command, was broken too: it filtered
out `StorageFormat::Inline` rows, so for any small series the UNION built
`FROM ()` and failed with a SQL parse error.

Nothing depended on any of it -- no testsuite coverage, no mention in
cli-reference.md. Removing it is a smaller change than making it correct, and
leaves the collapse path simpler than it was.

Removed: both CLI commands, TemporalFilteredListingTable (create_table_provider
now returns the ListingTable directly), the dead tlogfs filter module, the
min_override/max_override columns and their accessors,
get_temporal_overrides_for_node_id, the override branch of
set_extended_attributes, inherited_overrides, and
PersistenceLayer::get_temporal_bounds with its five implementations.

Kept `parse_timestamp_seconds`, moved into export.rs, its only remaining
consumer. `min_event_time`/`max_event_time` and `infer_temporal_bounds` are
unrelated series machinery and stay.

Existing ponds keep min_override/max_override as orphaned Delta columns; the
schema migration only adds columns, never drops them. Verified by building a
pond with the previous binary and then querying, listing, describing, fsck'ing,
and appending to it with this one. steward's NARROW_META_SQL named both columns
explicitly and had to be updated, or every write failed.

Copilot-Session: 84fa7705-d43d-4209-be99-854eba554370
