+++ crates/cmd/src/commands/mod.rs
@@ -27,7 +27,6 @@ pub mod run;
 pub mod run_summary;
 pub mod show;
 pub mod status;
-pub mod temporal;
 pub mod tlog;
 pub mod verify;
 
@@ -56,6 +55,5 @@ pub use restore::restore_command;
 pub use run::run_command;
 pub use show::show_command;
 pub use status::status_command;
-pub use temporal::{detect_overlaps_command, set_temporal_bounds_command};
 pub use tlog::{tlog_show_command, tlog_verify_command};
 pub use verify::verify_command;
