+++ crates/provider/src/transform/mod.rs
@@ -5,4 +5,3 @@
 pub mod column_rename;
 pub mod null_padding;
 pub mod scope_prefix;
-pub mod temporal_filter;
