+++ crates/tlogfs/src/query/mod.rs
@@ -3,6 +3,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod sql_executor;
-pub mod temporal_filter;
 
 pub use sql_executor::{execute_sql_on_file, get_file_schema};
