+++ crates/tlogfs/src/tests/mod.rs
@@ -918,326 +918,6 @@ async fn test_streaming_async_reader_small_file() -> Result<(), Box<dyn std::err
     Ok(())
 }
 
-/// Test temporal bounds functionality on file series
-///
-/// This test creates a file series with multiple versions, verifies count queries,
-/// sets temporal bounds, and validates that the bounds correctly filter the data.
-#[tokio::test]
-async fn test_temporal_bounds_on_file_series() -> Result<(), Box<dyn std::error::Error>> {
-    use crate::query::execute_sql_on_file;
-    use crate::schema::watertown;
-    use futures::stream::StreamExt;
-    use std::collections::HashMap;
-
-    debug!("=== Starting Temporal Bounds Test ===");
-
-    let store_path = test_dir();
-
-    // Create TLogFS persistence layer
-    debug!("Creating persistence layer...");
-    let mut persistence = OpLogPersistence::create_test(&store_path).await?;
-
-    let series_path = "test/temporal_series.series";
-
-    // Transaction 1: Create file series with all data points (simulating the combined data from multiple versions)
-    debug!(
-        "Creating file series with points at T=1,10,11,12,13,14,15,16,17,18,19,50 (12 points total)..."
-    );
-    {
-        let tx = persistence.begin_test().await?;
-        let wd = tx.root().await?;
-
-        // Create test directory if it doesn't exist
-        if !wd.exists(std::path::Path::new("test")).await {
-            _ = wd.create_dir_path("test").await?;
-        }
-
-        // Create batch with timestamps as proper Unix epoch seconds (not milliseconds)
-        // T=1,10,11,12,13,14,15,16,17,18,19,50 (12 points total)
-        use arrow_array::{Float64Array, RecordBatch, StringArray, TimestampSecondArray};
-        use arrow_schema::{DataType, Field, Schema, TimeUnit};
-        use std::sync::Arc;
-
-        let timestamp_data = TimestampSecondArray::from(vec![
-            Some(1_i64),
-            Some(10_i64),
-            Some(11_i64),
-            Some(12_i64),
-            Some(13_i64),
-            Some(14_i64),
-            Some(15_i64),
-            Some(16_i64),
-            Some(17_i64),
-            Some(18_i64),
-            Some(19_i64),
-            Some(50_i64),
-        ])
-        .with_timezone("+00:00");
-
-        let value_data = Float64Array::from(vec![
-            100.0_f64, 110.0_f64, 111.0_f64, 112.0_f64, 113.0_f64, 114.0_f64, 115.0_f64, 116.0_f64,
-            117.0_f64, 118.0_f64, 119.0_f64, 150.0_f64,
-        ]);
-
-        let sensor_data = StringArray::from(vec![
-            "sensor_001",
-            "sensor_001",
-            "sensor_001",
-            "sensor_001",
-            "sensor_001",
-            "sensor_001",
-            "sensor_001",
-            "sensor_001",
-            "sensor_001",
-            "sensor_001",
-            "sensor_001",
-            "sensor_001",
-        ]);
-
-        let schema = Arc::new(Schema::new(vec![
-            Field::new(
-                "timestamp",
-                DataType::Timestamp(TimeUnit::Second, Some("+00:00".into())),
-                false,
-            ),
-            Field::new("value", DataType::Float64, false),
-            Field::new("sensor_id", DataType::Utf8, false),
-        ]));
-
-        let batch = RecordBatch::try_new(
-            schema,
-            vec![
-                Arc::new(timestamp_data),
-                Arc::new(value_data),
-                Arc::new(sensor_data),
-            ],
-        )?;
-
-        let (min_time, max_time) = wd
-            .create_series_from_batch(series_path, &batch, Some("timestamp"))
-            .await?;
-        debug!("Series created. Time range: {} to {}", min_time, max_time);
-
-        tx.commit_test().await?;
-    }
-
-    // Transaction 3: Query the series before setting temporal bounds - should return 12 rows
-    debug!("Querying series before temporal bounds...");
-    {
-        let tx = persistence.begin_test().await?;
-        let wd = tx.root().await?;
-
-        let mut result_stream = execute_sql_on_file(
-            &wd,
-            series_path,
-            "SELECT COUNT(*) as row_count FROM source",
-            &tx.state()?.as_provider_context(),
-        )
-        .await?;
-
-        let mut total_count = 0_i64;
-        while let Some(batch_result) = result_stream.next().await {
-            let batch = batch_result?;
-            if batch.num_rows() > 0 {
-                let count_array = batch
-                    .column(0)
-                    .as_any()
-                    .downcast_ref::<arrow_array::Int64Array>()
-                    .expect("COUNT(*) should return Int64Array");
-                total_count = count_array.value(0);
-            }
-        }
-
-        debug!("Count before temporal bounds: {}", total_count);
-        assert_eq!(
-            total_count, 12,
-            "Should have 12 total rows before temporal bounds"
-        );
-
-        // Don't commit this read-only transaction
-    }
-
-    // Transaction 4: Set temporal bounds to [10000, 20000] ms using empty version approach
-    debug!("Setting temporal bounds to [10000, 20000] ms (10-20 seconds) using empty version...");
-    {
-        let tx = persistence.begin_test().await?;
-        let wd = tx.root().await?;
-
-        // Create an empty version with temporal bounds (following CLI pattern)
-        // Use TinyFS async writer to add empty version to existing file
-        let mut writer = wd
-            .async_writer_path_with_type(series_path, tinyfs::EntryType::TablePhysicalSeries)
-            .await?;
-
-        // Write empty content to create a pending record
-        use tokio::io::AsyncWriteExt;
-        writer.write_all(&[]).await?;
-        writer.flush().await?;
-        writer.shutdown().await?;
-
-        debug!("Created empty version for metadata");
-
-        // Set temporal overrides using extended attributes on the path
-        // Note: bounds need to be in milliseconds since that's the internal representation
-        let mut attributes = HashMap::new();
-        _ = attributes.insert(
-            watertown::MIN_TEMPORAL_OVERRIDE.to_string(),
-            "10000".to_string(),
-        ); // 10 seconds = 10000 ms
-        _ = attributes.insert(
-            watertown::MAX_TEMPORAL_OVERRIDE.to_string(),
-            "20000".to_string(),
-        ); // 20 seconds = 20000 ms
-
-        wd.set_extended_attributes(series_path, attributes).await?;
-
-        debug!("Temporal bounds set to [10000, 20000] milliseconds (10-20 seconds)");
-        tx.commit_test().await?;
-    }
-
-    // Transaction 5: Query the series after setting temporal bounds - should return 10 rows
-    debug!("Querying series after temporal bounds...");
-    {
-        let tx = persistence.begin_test().await?;
-        let wd = tx.root().await?;
-
-        let mut result_stream = execute_sql_on_file(
-            &wd,
-            series_path,
-            "SELECT COUNT(*) as row_count FROM source",
-            &tx.state()?.as_provider_context(),
-        )
-        .await?;
-
-        let mut total_count = 0_i64;
-        while let Some(batch_result) = result_stream.next().await {
-            let batch = batch_result?;
-            if batch.num_rows() > 0 {
-                let count_array = batch
-                    .column(0)
-                    .as_any()
-                    .downcast_ref::<arrow_array::Int64Array>()
-                    .expect("COUNT(*) should return Int64Array");
-                total_count = count_array.value(0);
-            }
-        }
-
-        debug!("Count after temporal bounds: {}", total_count);
-        assert_eq!(
-            total_count, 10,
-            "Should have 10 rows after temporal bounds [10000, 20000] ms (excluding T=1s and T=50s)"
-        );
-
-        // Don't commit this read-only transaction
-    }
-
-    // Transaction 6: keep appending after the bounds were set, so the series
-    // has enough versions to be collapsible and the override row is no longer
-    // the highest version.  The new points sit OUTSIDE [10s, 20s], so if the
-    // bounds ever stop applying the count jumps above 10.
-    debug!("Appending further out-of-bounds versions after the temporal bounds...");
-    for (i, t) in [100_i64, 200, 300, 400].into_iter().enumerate() {
-        let tx = persistence.begin_test().await?;
-        let wd = tx.root().await?;
-
-        use arrow_array::{Float64Array, RecordBatch, StringArray, TimestampSecondArray};
-        use arrow_schema::{DataType, Field, Schema, TimeUnit};
-        use std::sync::Arc;
-
-        let schema = Arc::new(Schema::new(vec![
-            Field::new(
-                "timestamp",
-                DataType::Timestamp(TimeUnit::Second, Some("+00:00".into())),
-                false,
-            ),
-            Field::new("value", DataType::Float64, false),
-            Field::new("sensor_id", DataType::Utf8, false),
-        ]));
-        let batch = RecordBatch::try_new(
-            schema,
-            vec![
-                Arc::new(TimestampSecondArray::from(vec![Some(t)]).with_timezone("+00:00")),
-                Arc::new(Float64Array::from(vec![900.0_f64 + i as f64])),
-                Arc::new(StringArray::from(vec!["sensor_001"])),
-            ],
-        )?;
-        _ = wd
-            .write_series_from_batch(series_path, &batch, Some("timestamp"))
-            .await?;
-        tx.commit_test().await?;
-    }
-
-    // Transaction 7: Collapse the series, then re-query.  The bounds must
-    // survive.
-    //
-    // Regression: a collapse output takes a fresh HIGHEST version while
-    // standing for content mid-stream, and merged rows never carry temporal
-    // overrides.  Resolving the override by `max_by_key(version)` therefore
-    // picked the merged run, found no override, and the caller fell back to
-    // (i64::MIN, i64::MAX) -- silently re-admitting the T=1s and T=50s rows
-    // that `set-temporal-bounds` exists to exclude, plus the new ones above.
-    debug!("Collapsing the series and re-checking temporal bounds...");
-    {
-        let tx = persistence.begin_test().await?;
-        {
-            let state = tx.state()?;
-            let candidates = state.list_collapsible_series(2).await?;
-            assert!(
-                !candidates.is_empty(),
-                "series should be collapsible after the appends above"
-            );
-            for id in candidates {
-                let stats = state.collapse_table_series(id, 2).await?;
-                assert!(stats.collapsed, "collapse should have merged a window");
-            }
-        }
-        tx.commit_test().await?;
-    }
-
-    {
-        let tx = persistence.begin_test().await?;
-        let wd = tx.root().await?;
-
-        let mut result_stream = execute_sql_on_file(
-            &wd,
-            series_path,
-            "SELECT COUNT(*) as row_count FROM source",
-            &tx.state()?.as_provider_context(),
-        )
-        .await?;
-
-        let mut total_count = 0_i64;
-        while let Some(batch_result) = result_stream.next().await {
-            let batch = batch_result?;
-            if batch.num_rows() > 0 {
-                let count_array = batch
-                    .column(0)
-                    .as_any()
-                    .downcast_ref::<arrow_array::Int64Array>()
-                    .expect("COUNT(*) should return Int64Array");
-                total_count = count_array.value(0);
-            }
-        }
-
-        debug!("Count after collapse: {}", total_count);
-        assert_eq!(
-            total_count, 10,
-            "temporal bounds must survive a collapse; got {total_count} rows, which means \
-             the override was not found and filtering silently reverted to unbounded"
-        );
-    }
-
-    debug!("SUCCESS: Temporal bounds test completed");
-    debug!(
-        "  - Created file series with data points at T=1,10,11,12,13,14,15,16,17,18,19,50 seconds"
-    );
-    debug!("  - Verified initial count of 12 rows");
-    debug!("  - Set temporal bounds to [10000, 20000] ms (10-20 seconds)");
-    debug!("  - Verified filtered count of 10 rows (excluded T=1s and T=50s)");
-
-    Ok(())
-}
-
 /// Test to reproduce unnecessary directory updates when appending to file:series
 ///
 /// This test writes multiple versions to the same file:series and checks
